requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
chromadb
gitpython
python-dotenv
pydantic
openai
